# api_document.json
{
    "openapi": "3.0.0",
    "paths": {
        "/auth/login": {
            "post": {
                "operationId": "login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TransformResponseDto"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/LoginResponseDto"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/auth/refresh-token": {
            "post": {
                "operationId": "jwtRefresh",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequestDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TransformResponseDto"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/RefreshTokenResponseDto"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Authentication"
                ]
            }
        },
...
    },
    "info": {
        "title": "ADMIN API",
        "description": "This api is used for managing mutiple betting pages",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
...
}